refactor: organize validation rules into domain-specific structure
- Removed 13 duplicate flat validation rules that had better domain-organized equivalents
- Updated all old Livewire forms to use new domain-organized rules
- Consolidated duplicate Stables rules (HasMinimumAmountOfMembers → HasMinimumMembers)
- Moved generic activation rule to Shared folder as CanChangeActivationDate
- Established clean domain-organized structure following architectural patterns
- All rules now properly organized by domain (Events/, Matches/, Referees/, etc.)
- Maintained backward compatibility by updating all imports and references
