+++ app/Livewire/Managers/ManagerForm.php
@@ -6,7 +6,7 @@ namespace App\Livewire\Managers;
 
 use App\Livewire\Base\LivewireBaseForm;
 use App\Models\Managers\Manager;
-use App\Rules\EmploymentStartDateCanBeChanged;
+use App\Rules\Shared\CanChangeEmploymentDate;
 use Illuminate\Support\Carbon;
 
 /**
@@ -48,14 +48,14 @@ class ManagerForm extends LivewireBaseForm
     }
 
     /**
-     * @return array<string, list<EmploymentStartDateCanBeChanged|string>>
+     * @return array<string, list<CanChangeEmploymentDate|string>>
      */
     protected function rules(): array
     {
         return [
             'first_name' => ['required', 'string', 'max:255'],
             'last_name' => ['required', 'string', 'max:255'],
-            'start_date' => ['nullable', 'date', new EmploymentStartDateCanBeChanged($this->formModel)],
+            'start_date' => ['nullable', 'date', new CanChangeEmploymentDate($this->formModel)],
         ];
     }
