--- app/Rules/Stables/HasMinimumAmountOfMembers.php
@@ -1,42 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Rules\Stables;
-
-use App\Models\Stables\Stable;
-use App\Models\TagTeams\TagTeam;
-use App\Models\Wrestlers\Wrestler;
-use Closure;
-use Illuminate\Contracts\Validation\ValidationRule;
-use Illuminate\Support\Collection;
-
-class HasMinimumAmountOfMembers implements ValidationRule
-{
-    /**
-     * @param  Collection<int, Wrestler>  $wrestlers
-     * @param  Collection<int, TagTeam>  $tagTeams
-     */
-    public function __construct(
-        private Stable $stable,
-        private readonly Collection $wrestlers,
-        private readonly Collection $tagTeams
-    ) {}
-
-    /**
-     * Determine if the validation rule passes.
-     */
-    public function validate(string $attribute, mixed $value, Closure $fail): void
-    {
-        if ($this->stable->isCurrentlyActivated()) {
-            $tagTeamsCountFromRequest = $this->tagTeams->count();
-            $wrestlersCountFromRequest = $this->wrestlers->count();
-
-            $tagTeamMembersCount = $tagTeamsCountFromRequest * 2;
-
-            if ($tagTeamMembersCount + $wrestlersCountFromRequest < Stable::MIN_MEMBERS_COUNT) {
-                $fail("{$this->stable->name} is currently activated and the activation date cannot be changed.");
-            }
-        }
-    }
-}
