--- app/Rules/TitleMustBeActive.php
@@ -1,25 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Rules;
-
-use App\Models\Titles\Title;
-use Closure;
-use Illuminate\Contracts\Validation\ValidationRule;
-
-class TitleMustBeActive implements ValidationRule
-{
-    /**
-     * Determine if the validation rule passes.
-     */
-    public function validate(string $attribute, mixed $value, Closure $fail): void
-    {
-        /** @var Title $title */
-        $title = Title::query()->whereKey($value)->sole();
-
-        if (! $title->isCurrentlyActivated()) {
-            $fail('This title is not active and cannot be added to the match.');
-        }
-    }
-}
