--- app/Rules/WrestlerCanJoinNewStable.php
@@ -1,35 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Rules;
-
-use App\Models\Wrestlers\Wrestler;
-use Closure;
-use Illuminate\Contracts\Validation\ValidationRule;
-use Illuminate\Support\Collection;
-
-class WrestlerCanJoinNewStable implements ValidationRule
-{
-    /**
-     * @param  Collection<int, int>  $tagTeamIds
-     */
-    public function __construct(private Collection $tagTeamIds) {}
-
-    /**
-     * Determine if the validation rule passes.
-     */
-    public function validate(string $attribute, mixed $value, Closure $fail): void
-    {
-        /** @var Wrestler $wrestler */
-        $wrestler = Wrestler::with(['currentStable'])->find($value);
-
-        if ($wrestler->currentStable()->getParent()->exists()) {
-            $fail('This wrestler is already a member of a stable.');
-        }
-
-        if ($this->tagTeamIds->isNotEmpty() && (! is_null($wrestler->currentTagTeam) && $this->tagTeamIds->contains($wrestler->currentTagTeam->id))) {
-            $fail('This wrestler is already a member of a stable.');
-        }
-    }
-}
