--- app/Rules/WrestlerMustBeBookable.php
@@ -1,27 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Rules;
-
-use App\Models\Wrestlers\Wrestler;
-use Closure;
-use Illuminate\Contracts\Validation\ValidationRule;
-
-class WrestlerMustBeBookable implements ValidationRule
-{
-    /**
-     * Run the validation rule.
-     *
-     * @param  Closure(string): \Illuminate\Translation\PotentiallyTranslatedString  $fail
-     */
-    public function validate(string $attribute, mixed $value, Closure $fail): void
-    {
-        /** @var Wrestler $wrestler */
-        $wrestler = Wrestler::find($value);
-
-        if (! $wrestler->isBookable()) {
-            $fail('This wrestler is not bookable and cannot be added to the match.');
-        }
-    }
-}
